Modularised table component
